# expense-management-app/API.yaml
openapi: 3.0.0
info:
  title: Expense Management API
  version: 1.0.0
  description: Full-stack expense management API with payment processing and AI insights
  contact:
    name: Support
    url: https://github.com/yourusername/expense-management-app
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
paths:
...
  /summary/generate:
    post:
      summary: Generate expense summary with AI insights
      tags: [Summary]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSummary'
  /summary:
    get:
      summary: Get all summaries for user
      tags: [Summary]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of summaries
          content:
            application/json:
              schema:
                type: object
                properties:
                  summaries:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExpenseSummary'
...
components:
  schemas:
...
    ExpenseSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        total_expenses:
          type: number
          format: decimal
        category_breakdown:
          type: object
        ai_summary:
          type: string
        insights:
          type: object
        created_at:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT